requirements.txt
biopython>=1.83
pytest>=8.0
